refactor: split sample browser render logic
