+++ src/egui_app/ui.rs
@@ -19,7 +19,8 @@ mod map_view;
 mod platform;
 mod progress_overlay;
 mod sample_browser_filter;
-mod sample_browser_panel;
+mod sample_browser_interaction;
+mod sample_browser_render;
 mod sample_browser_rename;
 mod sample_menus;
 mod sources_panel;
